feat: move settings and stores under kernel
