+++ apps/web/kernel/domain/index.ts
@@ -1,3 +1,4 @@
 export * from './entities';
 export * from './repositories';
+export * from './stores';
 export * from './value-object';
