+++ apps/web/kernel/index.ts
@@ -2,3 +2,4 @@ export * from './application';
 export * from './domain';
 export * from './infrastructure';
 export * from './init';
+export * from './presentation';
