+++ apps/web/kernel/presentation/settings/ExperienceSettings.tsx
@@ -9,7 +9,7 @@ import {
 import { useTranslations } from 'next-intl';
 import { FunctionComponent } from 'react';
 
-import { useExperienceFlagsStore } from '@/state/experience_flags';
+import { useExperienceFlagsStore } from '@/kernel';
 
 type ExperienceSettingsProps = {
   className?: string;
