+++ apps/web/kernel/presentation/settings/index.ts
@@ -1,0 +1,5 @@
+export { default as ExperienceSettings } from './ExperienceSettings';
+export { default as PreferencesSettings } from './PreferencesSettings';
+export { default as RuntimeSettings } from './RuntimeSettings';
+export { default as UserGrantsSettings } from './UserGrantsSettings';
+export { default as MandatoryExperienceSettings } from './MandatoryExperienceSettings';
