+++ apps/web/src/components/CookieConsent.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from 'next-intl';
 import { FunctionComponent } from 'react';
 
 import { Link } from '@/i18n/navigation';
-import { useUserGrantsStore } from '@/state/user_grants';
+import { useUserGrantsStore } from '@/kernel';
 
 const CookieConsent: FunctionComponent = () => {
   const t = useTranslations();
