+++ apps/web/src/features/browser_core/hooks/useDisableNavigationPop.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { useExperienceFlagsStore } from '@/state/experience_flags';
+import { useExperienceFlagsStore } from '@/kernel';
 
 // This NEEDS user interaction first, otherwise it won't work at all.
 // Also, this can't really have a cleanup function either.
