--- apps/web/src/features/settings/index.ts
@@ -1,5 +1,0 @@
-export { default as ExperienceSettings } from './components/ExperienceSettings';
-export { default as PreferencesSettings } from './components/PreferencesSettings';
-export { default as RuntimeSettings } from './components/RuntimeSettings';
-export { default as UserGrantsSettings } from './components/UserGrantsSettings';
-export { default as MandatoryExperienceSettings } from './components/MandatoryExperienceSettings';
